Modified csvreader & created test file
Modified csvreader to look for a command line to point to a file to read, it defaults to Magic10Normal5Items.csv if no file is specified.

Also added a test file Items10GarbageItems.csv. This was for testing, but is a partial file that will be used as part of a larger more complicated test file in the future.