Cleaned up csv reader, started the garbage item test case generator.
Cleaned up the CSV reader. Included a description of the code. Added documentation comments.
Rebuilt the program to be a function, and a main to call it for testing.

Started work on a “garbage item” csv generator. Currently I have copied the 10 item list created yesterday for this to the generator directory and renamed it as “ItemList.csv”. I have also created a script that will be default use this list to update or create a config file that has the name of the list file and the number of items in the file. It is possible to tell it to use a different file as a base. Once the file is changed the script can be rerun and it will use the new item list instead of the default.