Added Configs To ConfigTool, Headers to Item list and ItemListConfigUpdater fix
Added 10 Configs and 1 setting to ConfigTool. Including csv formattting "errors" and Invalid Item entry types. It also includes the option to have the entries in the test file randomized.

To match the other item files column headers were added to ItemList.csv

ItemListconfigUpdater.py was altered to read a list with headers, update a comment, and added the a count to the item config, to allow user to limit the number of real items used in the test files.