Config Tool behavior and formatting correction, ItemListConfigUpdater formatting fix
Config Tool behaviour and fomatting fix
The tool incorrectly did nothing when a config file was found missing, now it informs the user that the file was missing. The comment was corrected to reflect this.
The formatting was also changed to use only tabs, instead of a mix of tabs and spaces.

ItemListConfigUpdater
The formatting was changed to use only tabs, instead of a mix of tabs and spaces.